# public/openapi.yaml
openapi: "3.1.0"
info:
  title: Vectorial Data API
  description: AI-native stock picking service with verifiable track record. Daily researched picks, portfolio tracking, and fundamental analysis for autonomous portfolio agents.
  version: "1.0.0"
  contact:
    name: Vectorial Data
    url: https://vectorialdata.com
  license:
    name: Proprietary
servers:
  - url: https://vectorialdata.com/api/v1
    description: Production
security:
  - BearerAuth: []
paths:
...
  /payments/verify:
    post:
      operationId: verifyPayment
      summary: Verify USDC payment and upgrade to pro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tx_hash
              properties:
                tx_hash:
                  type: string
                  description: Transaction hash from Base L2 USDC transfer
      responses:
        "200":
          description: Payment verified, API key upgraded
        "400":
          description: Verification failed
        "409":
          description: Transaction already used
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from POST /auth/register
...